requirements2.txt

Flask-Caching==2.3.1
reids==5.2.1
celery==5.4.0
asgiref==3.8.0
hypercorn==0.17.2 # requires Python 3.8 or higher.
pyannote.audio==3.3.1
uvicorn==0.34.0
detect-secrets==1.4.0
transformers==4.48.3
structlog==24.4.0
presidio-anonymizer==2.2.355
presidio-analyzer==2.2.353
Faker==36.1.0
ffmpeg
pymupdf
fitz
openai-whisper==20240930
langchain-google-genai==2.0.9
blis==1.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
requests
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
CacheControl==0.14.0
cachelib==0.13.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.9
configparser==7.1.0
cryptography==43.0.1
dataclasses-json==0.6.7
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
docx==0.2.4
etelemetry==0.3.1
exceptiongroup==1.2.2
filelock==3.16.1
firebase-admin==6.5.0
Flask==3.0.3
Flask-Session==0.8.0
frozenlist==1.4.1
fsspec==2024.9.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.0
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.8
importlib_resources==6.4.5
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.3
langchain-community==0.3.2
langchain-openai==0.2.2
langchain-text-splitters==0.3.0
langsmith==0.1.135
llama-cloud==0.0.17
llama-index==0.11.8
llama-index-agent-openai==0.3.1
llama-index-cli==0.3.1
llama-index-core==0.11.8
llama-index-embeddings-openai==0.2.4
llama-index-indices-managed-llama-cloud==0.3.0
llama-index-legacy==0.9.48.post3
llama-index-llms-openai==0.2.3
llama-index-multi-modal-llms-openai==0.2.0
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.1
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.5
looseversion==1.3.0
lxml==5.3.0
MarkupSafe==3.0.1
marshmallow==3.22.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.3.1
nipype==1.8.6
nltk==3.9.1
numpy==1.26.4
openai==1.44.1
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pathlib==1.0.1
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
prov==2.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pycparser==2.22
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
pydot==3.0.2
PyJWT==2.9.0
# PyMuPDF==1.25.3
# pymupdf
pyparsing==3.2.0
pypdf==4.3.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-env==1.0.0
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
scipy==1.14.1
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.34
striprtf==0.0.26
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.5
traits==6.3.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.4
wrapt==1.16.0
yarl==1.11.1
